Moving number packing helpers into own file.
