+++ src/github.com/acook/blacklight/vm.go
@@ -46,23 +46,3 @@ func run_vm(bc []byte) {
 		}
 	}
 }
-
-func Varint32(buf []byte) int32 {
-	ux := binary.BigEndian.Uint32(buf)
-
-	x := int32(ux >> 1)
-	if ux&1 != 0 {
-		x = ^x
-	}
-	return x
-}
-
-func Varint64(buf []byte) int64 {
-	ux := binary.BigEndian.Uint64(buf)
-
-	x := int64(ux >> 1)
-	if ux&1 != 0 {
-		x = ^x
-	}
-	return x
-}
